+++ sigopt/compat.py
@@ -10,14 +10,3 @@ except ImportError:
       'No json library installed.'
       ' Try running `pip install simplejson` to install a compatible json library.'
     ) from ie
-
-
-try:
-  import xgboost
-  from xgboost import Booster, DMatrix
-  XGBOOST_INSTALLED = True
-except ImportError:
-  xgboost = False
-  Booster = object
-  DMatrix = object
-  XGBOOST_INSTALLED = False
